Organização do frontend + Adição do React Router
