+++ src/App.jsx
@@ -1,43 +1,15 @@
-import logo from './logo.svg';
-import './App.css';
+import {
+    BrowserRouter as Router,
+    Route,
+    Switch
+} from 'react-router-dom';
 
-import { useState, useEffect } from 'react';
+import Home from './views/home';
 
-import axios from 'axios';
-
-function App() {
-    const [message, setMessage] = useState('Carregando...');
-
-    useEffect(() => {
-        const fetchData = async () => {
-            const { data } = await axios.get('/api/hello');
-
-            setMessage(data.message);
-        }
-
-        setInterval(() => {
-            fetchData();
-        }, 1000);
-    }, []);
-
-    return (
-        <div className="App">
-            <header className="App-header">
-                <img src={logo} className="App-logo" alt="logo" />
-                <p>
-                    { message } <br />
-                </p>
-                <a
-                    className="App-link"
-                    href="https://reactjs.org"
-                    target="_blank"
-                    rel="noopener noreferrer"
-                >
-                    Learn React
-                </a>
-            </header>
-        </div>
-    );
-}
+const App = () => <Router>
+    <Switch>
+        <Route path="/" exact component={Home} />
+    </Switch>
+</Router>
 
 export default App;
